Remove internal z commands from 2605.svg path.
picosvg was parsing the path incorrectly when the `z` command (which returns the pointer to the start of the path) showed up in the middle of the path string. 
This version instead explicitly states the starting coords to move to. It should be functionally equivalent, but won't trip up picosvg.