Delete 3000.svg
This ideographic space was causing the font to fail to render properly on Firefox (when compiled with nanoemoji 0.8.1). 
I vaguely suspect this had something to do with with the blank glyph messing up the colr table in a non-catastrophic way. 🤷 
Anywho.
This is one change to this branch that they probably won't want to merge back upstream. The rest are pretty straight improvements.